+++ core/src/main/java/org/apache/calcite/avatica/remote/looker/LookerRemoteMeta.java
@@ -18,8 +18,6 @@ package org.apache.calcite.avatica.remote.looker;
 
 import org.apache.calcite.avatica.AvaticaConnection;
 import org.apache.calcite.avatica.AvaticaStatement;
-import org.apache.calcite.avatica.ColumnMetaData;
-import org.apache.calcite.avatica.ColumnMetaData.Rep;
 import org.apache.calcite.avatica.Meta;
 import org.apache.calcite.avatica.MissingResultsException;
 import org.apache.calcite.avatica.NoSuchStatementException;
@@ -30,8 +28,6 @@ import org.apache.calcite.avatica.remote.Service.PrepareAndExecuteRequest;
 import org.apache.calcite.avatica.remote.Service.PrepareRequest;
 import org.apache.calcite.avatica.remote.TypedValue;
 
-import com.fasterxml.jackson.core.JsonFactory;
-import com.fasterxml.jackson.core.JsonParser;
 import com.looker.rtl.AuthSession;
 import com.looker.rtl.Transport;
 import com.looker.sdk.LookerSDK;
@@ -42,8 +38,6 @@ import java.net.URL;
 import java.security.KeyManagementException;
 import java.security.NoSuchAlgorithmException;
 import java.sql.SQLException;
-import java.sql.Types;
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.ArrayBlockingQueue;
@@ -61,6 +55,7 @@ import javax.net.ssl.X509TrustManager;
  * from the Looker SDK.
  */
 public class LookerRemoteMeta extends RemoteMeta implements Meta {
+
   private final LookerRemoteService lookerService;
 
   public LookerRemoteMeta(AvaticaConnection connection, Service service) {
@@ -69,12 +64,6 @@ public class LookerRemoteMeta extends RemoteMeta implements Meta {
     assert LookerRemoteService.class.isAssignableFrom(service.getClass());
     lookerService = (LookerRemoteService) service;
   }
-
-  /**
-   * Constants used in JSON parsing
-   */
-  private static final String ROWS_KEY = "rows";
-  private static final String VALUE_KEY = "value";
 
   /**
    * Default queue size. Could probably be more or less. 10 chosen for now.
@@ -93,124 +82,8 @@ public class LookerRemoteMeta extends RemoteMeta implements Meta {
    * {@code FrameEnvelopes}s that belong to a running statement. See {@link #prepareStreamingThread}
    * for more details.
    */
-  private final ConcurrentMap<StatementHandle, BlockingQueue<FrameEnvelope>> stmtQueueMap =
-      new ConcurrentHashMap();
-
-  /**
-   * Wrapper for either a {@link Frame} or {@link Exception}. Allows for {@link #stmtQueueMap} to
-   * hold complete Frames and present exceptions when consumers are ready to encounter them.
-   */
-  private class FrameEnvelope {
-
-    final Frame frame;
-    final Exception exception;
-
-    FrameEnvelope(/*@Nullable*/ Frame frame, /*@Nullable*/ Exception exception) {
-      this.frame = frame;
-      this.exception = exception;
-    }
-
-    boolean hasException() {
-      return this.exception != null;
-    }
-  }
-
-  private FrameEnvelope makeFrame(long offset, boolean done, Iterable<Object> rows) {
-    Frame frame = new Frame(offset, done, rows);
-    return new FrameEnvelope(frame, null);
-  }
-
-  private FrameEnvelope makeException(Exception e) {
-    return new FrameEnvelope(null, e);
-  }
-
-  /**
-   * Calls the correct method to read the current value on the stream. The {@code get} methods do
-   * not advance the current token, so they can be called multiple times without changing the state
-   * of the parser.
-   *
-   * @param columnMetaData the {@link ColumnMetaData} for this value. It is important to use the
-   *     {@link Rep} rather than the type name since Avatica represents most datetime values as
-   *     milliseconds since epoch via {@code long}s or {@code int}s.
-   * @param parser a JsonParser whose current token is a value from the JSON response. Callers
-   *     must ensure that the parser is ready to consume a value token. This method does not change
-   *     the state of the parser.
-   * @return the parsed value.
-   */
-  static Object deserializeValue(JsonParser parser, ColumnMetaData columnMetaData)
-      throws IOException {
-    switch (columnMetaData.type.rep) {
-    case PRIMITIVE_BOOLEAN:
-    case BOOLEAN:
-      return parser.getBooleanValue();
-    case PRIMITIVE_BYTE:
-    case BYTE:
-      return parser.getByteValue();
-    case STRING:
-      return parser.getValueAsString();
-    case PRIMITIVE_SHORT:
-    case SHORT:
-      return parser.getShortValue();
-    case PRIMITIVE_INT:
-    case INTEGER:
-      return parser.getIntValue();
-    case PRIMITIVE_LONG:
-    case LONG:
-      return parser.getLongValue();
-    case PRIMITIVE_FLOAT:
-    case FLOAT:
-      return parser.getFloatValue();
-    case PRIMITIVE_DOUBLE:
-    case DOUBLE:
-      return parser.getDoubleValue();
-    case NUMBER:
-      // NUMBER is represented as BigDecimal
-      return parser.getDecimalValue();
-    // TODO: MEASURE types are appearing as Objects. This should have been solved by CALCITE-5549.
-    //  Be sure that the signature of a prepared query matches the metadata we see from JDBC.
-    case OBJECT:
-      switch (columnMetaData.type.id) {
-      case Types.INTEGER:
-        return parser.getIntValue();
-      case Types.BIGINT:
-        return parser.getBigIntegerValue();
-      case Types.DOUBLE:
-        return parser.getDoubleValue();
-      case Types.DECIMAL:
-      case Types.NUMERIC:
-        return parser.getDecimalValue();
-      }
-    default:
-      throw new IOException("Unable to parse " + columnMetaData.type.rep + " from stream!");
-    }
-  }
-
-  @Override
-  public Frame fetch(final StatementHandle h, final long offset, final int fetchMaxRowCount)
-      throws NoSuchStatementException, MissingResultsException {
-    // If this statement was initiated as a LookerFrame then it will have an entry in the queue map
-    if (stmtQueueMap.containsKey(h)) {
-      try {
-        BlockingQueue queue = stmtQueueMap.get(h);
-
-        // `take` blocks until there is an entry in the queue
-        FrameEnvelope nextEnvelope = (FrameEnvelope) queue.take();
-
-        // remove the statement from the map if it has an exception, or it is the last frame
-        if (nextEnvelope.hasException()) {
-          stmtQueueMap.remove(h);
-          throw new RuntimeException(nextEnvelope.exception);
-        } else if (nextEnvelope.frame.done) {
-          stmtQueueMap.remove(h);
-        }
-        return nextEnvelope.frame;
-      } catch (InterruptedException e) {
-        throw new RuntimeException(e);
-      }
-    }
-    // not a streaming query - default to RemoteMeta behavior
-    return super.fetch(h, offset, fetchMaxRowCount);
-  }
+  final ConcurrentMap<Integer, BlockingQueue<FrameEnvelope>> stmtQueueMap =
+      new ConcurrentHashMap<>();
 
   /**
    * An initially empty frame specific to Looker result sets. {@link #sqlInterfaceQueryId} is used
@@ -247,21 +120,17 @@ public class LookerRemoteMeta extends RemoteMeta implements Meta {
    */
   private void trustAllHosts(HttpsURLConnection connection) {
     // Create a trust manager that does not validate certificate chains
-    TrustManager[] trustAllCerts = new TrustManager[]{
-        new X509TrustManager() {
-          public java.security.cert.X509Certificate[] getAcceptedIssuers() {
-            return null;
-          }
+    TrustManager[] trustAllCerts = new TrustManager[]{new X509TrustManager() {
+      public java.security.cert.X509Certificate[] getAcceptedIssuers() {
+        return null;
+      }
 
-          public void checkClientTrusted(java.security.cert.X509Certificate[] certs,
-              String authType) {
-          }
+      public void checkClientTrusted(java.security.cert.X509Certificate[] certs, String authType) {
+      }
 
-          public void checkServerTrusted(java.security.cert.X509Certificate[] certs,
-              String authType) {
-          }
-        }
-    };
+      public void checkServerTrusted(java.security.cert.X509Certificate[] certs, String authType) {
+      }
+    }};
     // Create all-trusting host name verifier
     HostnameVerifier trustAllHostNames = (hostname, session) -> true;
     try {
@@ -317,87 +186,43 @@ public class LookerRemoteMeta extends RemoteMeta implements Meta {
     }
   }
 
-  private void seekToRows(JsonParser parser) throws IOException {
-    while (parser.nextToken() != null && !ROWS_KEY.equals(parser.currentName())) {
-      // move position to start of `rows`
-    }
-  }
-
-  private void seekToValue(JsonParser parser) throws IOException {
-    while (parser.nextToken() != null && !VALUE_KEY.equals(parser.currentName())) {
-      // seeking to `value` key for the field e.g. `"rows": [{"field_1": {"value": 123 }}]
-    }
-    // now move to the actual value
-    parser.nextToken();
-  }
-
-  private void putExceptionOrFail(BlockingQueue queue, Exception e) {
-    try {
-      // `put` blocks until there is room on the queue but needs a catch
-      queue.put(makeException(e));
-    } catch (InterruptedException ex) {
-      throw new RuntimeException(ex);
-    }
-  }
-
   /**
    * Prepares a thread to stream a query response into a series of {@link FrameEnvelope}s.
    */
   protected Thread prepareStreamingThread(String baseUrl, Signature signature, int fetchSize,
-      BlockingQueue frameQueue) {
-    Thread stream = new Thread(() -> {
+      BlockingQueue<FrameEnvelope> frameQueue) throws IOException {
+
+    InputStream in = makeRunQueryRequest(baseUrl);
+    LookerResponseParser parser = new LookerResponseParser(frameQueue);
+
+    return new Thread(() -> parser.parseStream(in, signature, fetchSize));
+  }
+
+  @Override
+  public Frame fetch(final StatementHandle h, final long offset, final int fetchMaxRowCount)
+      throws NoSuchStatementException, MissingResultsException {
+    // If this statement was initiated as a LookerFrame then it will have an entry in the queue map
+    if (stmtQueueMap.containsKey(h.id)) {
       try {
-        InputStream in = makeRunQueryRequest(baseUrl);
-        JsonParser parser = new JsonFactory().createParser(in);
+        BlockingQueue<FrameEnvelope> queue = stmtQueueMap.get(h.id);
 
-        int currentOffset = 0;
+        // `take` blocks until there is an entry in the queue
+        FrameEnvelope nextEnvelope = queue.take();
 
-        while (parser.nextToken() != null) {
-          if (currentOffset == 0) {
-            // TODO: Handle `metadata`. We are currently ignoring it and seeking to `rows` array
-            seekToRows(parser);
-          }
-
-          int rowsRead = 0;
-          List<Object> rows = new ArrayList<>();
-
-          while (rowsRead < fetchSize) {
-            List<Object> columnValues = new ArrayList<>();
-            // the signature should _always_ have the correct number of columns.
-            // if not, something went wrong during query preparation on the Looker instance.
-            for (int i = 0; i < signature.columns.size(); i++) {
-              seekToValue(parser);
-
-              if (parser.isClosed()) {
-                // the stream is closed - all rows should be accounted for
-                currentOffset += rowsRead;
-                frameQueue.put(makeFrame(currentOffset, /*done=*/true, rows));
-                return;
-              }
-
-              // add the value to the column list
-              columnValues.add(deserializeValue(parser, signature.columns.get(i)));
-            }
-
-            // Meta.CursorFactory#deduce will select an OBJECT cursor if there is only a single
-            // column in the signature. This is intended behavior. Since the rows of a frame are
-            // simply `Object` - we could get illegal casts from an ArrayList to the underlying Rep
-            // type if we don't discard the wrapping ArrayList for the single value here.
-            Object row = signature.columns.size() == 1 ? columnValues.get(0) : columnValues;
-
-            rows.add(row);
-            rowsRead++;
-          }
-          // we fetched the allowed number of rows so add the complete frame to the queue
-          currentOffset += rowsRead;
-          frameQueue.put(makeFrame(currentOffset, /*done=*/false, rows));
+        // remove the statement from the map if it has an exception, or it is the last frame
+        if (nextEnvelope.hasException()) {
+          stmtQueueMap.remove(h.id);
+          throw new RuntimeException(nextEnvelope.exception);
+        } else if (nextEnvelope.frame.done) {
+          stmtQueueMap.remove(h.id);
         }
-      } catch (Exception e) {
-        // enqueue all exceptions for the main thread to report
-        putExceptionOrFail(frameQueue, e);
+        return nextEnvelope.frame;
+      } catch (InterruptedException e) {
+        throw new RuntimeException(e);
       }
-    });
-    return stream;
+    }
+    // not a streaming query - default to RemoteMeta behavior
+    return super.fetch(h, offset, fetchMaxRowCount);
   }
 
   /**
@@ -421,14 +246,14 @@ public class LookerRemoteMeta extends RemoteMeta implements Meta {
         }
 
         // setup queue to place complete frames
-        BlockingQueue frameQueue = new ArrayBlockingQueue<FrameEnvelope>(DEFAULT_FRAME_QUEUE_SIZE);
+        BlockingQueue<FrameEnvelope> frameQueue = new ArrayBlockingQueue(DEFAULT_FRAME_QUEUE_SIZE);
 
         // update map so this statement is associated with a queue
-        stmtQueueMap.put(stmt.handle, frameQueue);
+        stmtQueueMap.put(stmt.handle.id, frameQueue);
 
         // init and start a new thread to stream from a Looker instance and populate the frameQueue
         prepareStreamingThread(url, signature, stmt.getFetchSize(), frameQueue).start();
-      } catch (SQLException | NoSuchStatementException e) {
+      } catch (SQLException | NoSuchStatementException | IOException e) {
         throw new RuntimeException(e);
       }
     }
