+++ core/src/main/java/org/apache/calcite/avatica/remote/looker/LookerRemoteService.java
@@ -17,6 +17,7 @@
 package org.apache.calcite.avatica.remote.looker;
 
 import org.apache.calcite.avatica.Meta.Signature;
+import org.apache.calcite.avatica.Meta.StatementHandle;
 import org.apache.calcite.avatica.remote.JsonService;
 import org.apache.calcite.avatica.remote.looker.LookerRemoteMeta.LookerFrame;
 
@@ -27,6 +28,9 @@ import com.looker.sdk.WriteSqlInterfaceQueryCreate;
 
 import java.io.IOException;
 import java.util.Arrays;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentMap;
+import java.util.concurrent.atomic.AtomicInteger;
 
 import static org.apache.calcite.avatica.remote.looker.LookerSdkFactory.safeSdkCall;
 
@@ -36,21 +40,69 @@ import static org.apache.calcite.avatica.remote.looker.LookerSdkFactory.safeSdkC
  */
 public class LookerRemoteService extends JsonService {
 
+  /**
+   * Keep track of next statement ID. This is similar to `CalciteMetaImpl` in that it only
+   * increments. Statements are scoped to connections, so it is unlikely we will ever encounter
+   * overflow. If we do, the client must make a new connection.
+   */
+  private final AtomicInteger statementCounter = new AtomicInteger(0);
+
+  /**
+   * These are statements that have been prepared but not yet executed. We need to know the query id
+   * to run on the Looker instance when a client decides to `execute` them. Multiple statements can
+   * run the same query, so it is important to index on the statement id rather than query id.
+   */
+  final ConcurrentMap<Integer, SqlQueryWithSignature> preparedStmtToQueryMap =
+      new ConcurrentHashMap<>();
+
+  /**
+   * The authenticated SDK used to communicate with a Looker instance.
+   */
   public LookerSDK sdk;
 
   void setSdk(LookerSDK sdk) {
     this.sdk = sdk;
   }
 
+  void checkSdk() {
+    assert null != sdk : "No authenticated SDK for this connection!";
+  }
+
+  private class SqlQueryWithSignature {
+
+    SqlInterfaceQuery query;
+    Signature signature;
+
+    SqlQueryWithSignature(SqlInterfaceQuery query) {
+      this.query = query;
+
+      try {
+        this.signature = JsonService.MAPPER.readValue(query.getSignature(), Signature.class);
+      } catch (IOException e) {
+        throw handle(e);
+      }
+    }
+  }
+
   /**
    * Helper method to create a {@link ExecuteResponse} for this request. Since we are using the
    * Looker SDK we need to create this response client side.
    */
-  ExecuteResponse lookerExecuteResponse(PrepareAndExecuteRequest request, Signature signature,
+  ExecuteResponse lookerExecuteResponse(String connectionId, int statementId, Signature signature,
       LookerFrame lookerFrame) {
-    ResultSetResponse rs = new ResultSetResponse(request.connectionId, request.statementId, false,
-        signature, lookerFrame, -1, null);
+    ResultSetResponse rs = new ResultSetResponse(connectionId, statementId, false, signature,
+        lookerFrame, -1, null);
     return new ExecuteResponse(Arrays.asList(new ResultSetResponse[]{rs}), false, null);
+  }
+
+  private SqlQueryWithSignature prepareQuery(String sql) {
+    checkSdk();
+    WriteSqlInterfaceQueryCreate queryRequest = new WriteSqlInterfaceQueryCreate(
+        sql, /*jdbcClient=*/true);
+    SqlInterfaceQuery preparedQuery = safeSdkCall(
+        () -> sdk.create_sql_interface_query(queryRequest));
+
+    return new SqlQueryWithSignature(preparedQuery);
   }
 
   /**
@@ -61,9 +113,56 @@ public class LookerRemoteService extends JsonService {
    */
   @Override
   public String apply(String request) {
-    assert null != sdk : "No authenticated SDK";
+    checkSdk();
     JdbcInterface response = safeSdkCall(() -> sdk.jdbc_interface(request));
+
     return response.getResults();
+  }
+
+  /**
+   * Handles CreateStatementRequest. We want to control the statement id since Looker instances do
+   * not keep track of the statement.
+   */
+  @Override
+  public CreateStatementResponse apply(CreateStatementRequest request) {
+    return new CreateStatementResponse(request.connectionId, statementCounter.getAndIncrement(),
+        /*rpcMetadata=*/ null);
+  }
+
+  /**
+   * Handles PrepareRequests by preparing a query via {@link LookerSDK#create_sql_query} whose
+   * response contains a query id. This id is used to execute the query via
+   * {@link LookerSDK#run_sql_query} with the 'json_bi' format.
+   *
+   * @param request the base Avatica request to convert into a Looker SDK call.
+   * @return a {@link PrepareResponse} containing a new {@link StatementHandle}.
+   */
+  @Override
+  public PrepareResponse apply(PrepareRequest request) {
+    checkSdk();
+    int currentStatementId = statementCounter.getAndIncrement();
+
+    SqlQueryWithSignature preparedQuery = prepareQuery(request.sql);
+    StatementHandle stmt = new StatementHandle(request.connectionId, currentStatementId,
+        preparedQuery.signature);
+    preparedStmtToQueryMap.put(currentStatementId, preparedQuery);
+
+    return new PrepareResponse(stmt, null);
+  }
+
+  /**
+   * Handles ExecuteRequests by setting up a {@link LookerFrame} to stream the response.
+   *
+   * @param request the base Avatica request. Used to locate the query to run in the statement map.
+   * @return a {@link ExecuteResponse} containing a prepared {@link LookerFrame}.
+   */
+  @Override
+  public ExecuteResponse apply(ExecuteRequest request) {
+    checkSdk();
+    SqlQueryWithSignature preparedQuery = preparedStmtToQueryMap.get(request.statementHandle.id);
+
+    return lookerExecuteResponse(request.statementHandle.connectionId, request.statementHandle.id,
+        preparedQuery.signature, LookerFrame.create(preparedQuery.query.getId()));
   }
 
   /**
@@ -76,19 +175,19 @@ public class LookerRemoteService extends JsonService {
    */
   @Override
   public ExecuteResponse apply(PrepareAndExecuteRequest request) {
-    assert null != sdk;
+    checkSdk();
+    SqlQueryWithSignature preparedQuery = prepareQuery(request.sql);
 
-    WriteSqlInterfaceQueryCreate queryRequest = new WriteSqlInterfaceQueryCreate(
-        request.sql, /*jdbcClient=*/true);
-    SqlInterfaceQuery preparedQuery = safeSdkCall(
-        () -> sdk.create_sql_interface_query(queryRequest));
+    return lookerExecuteResponse(request.connectionId, request.statementId, preparedQuery.signature,
+        LookerFrame.create(preparedQuery.query.getId()));
+  }
 
-    Signature signature;
-    try {
-      signature = JsonService.MAPPER.readValue(preparedQuery.getSignature(), Signature.class);
-    } catch (IOException e) {
-      throw handle(e);
-    }
-    return lookerExecuteResponse(request, signature, LookerFrame.create(preparedQuery.getId()));
+  /**
+   * If the statement is closed, clean up the prepared statement map
+   */
+  @Override
+  public CloseStatementResponse apply(CloseStatementRequest request) {
+    preparedStmtToQueryMap.remove(request.statementId);
+    return super.apply(request);
   }
 }
